+++ composables/useChain.ts
@@ -30,6 +30,9 @@ export default function () {
       : symbol
   })
 
+  const blockExplorer = computed(
+    () => chainProperties.value.blockExplorer ?? 'https://kusama.subscan.io/'
+  )
   return {
     decimals,
     unit,
@@ -37,5 +40,6 @@ export default function () {
     chainProperties,
     availableChains,
     chainSymbol,
+    blockExplorer,
   }
 }
