+++ locales/en.json
@@ -889,7 +889,8 @@
     "btnCopyPayment": "Copy Payment link",
     "congratsSupport": "Congratulations for supporting",
     "tweetDonation": "Tweet about your awesome donation",
-    "donationSentTo": "Your donation will be sent to:"
+    "donationSentTo": "Your donation will be sent to:",
+    "route": "Send {0} From {1} To {2}"
   },
   "preferences": {
     "basic": "basic user interface",
